requirements.txt
anyio==4.9.0
certifi==2025.6.15
cffi==1.17.1
commonx==0.6.38
curl-cffi==0.11.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmcomic==2.6.0
pillow==11.3.0
pycparser==2.22
pycryptodome==3.23.0
python-telegram-bot==22.2
PyYAML==6.0.2
sniffio==1.3.1
typing-extensions==4.14.0